refactor: organise core package
